Standardize on "open source" without a hyphen
There's a raging debate over whether a hyphen should be used in
"open-source." Generically, one would always include a hyphen in such a
compound term. But, as this discussion on the Wikipedia makes clear
(http://en.wikipedia.org/wiki/Talk:Open-source_software), the term in
practice tend strongly omits the hyphen, because there is no ambiguity
now that the phrase stands alone, cognitively.

In writing "open-source," we look a bit like we're writing
"to-morrow"—it smacks of old-fashioned. Hence standardizing on "open
source," given the likely audience for Project Open Data.